AFM horizontal or vertical profile line

create a profile data line for a particular position given in pixels; either a horizontal or a vertical line; this is also an easy way to reiterate through all lines of an image

AFM.getLine(
  obj,
  xPixel = NA,
  yPixel = NA,
  no = 1,
  dataOnly = FALSE,
  verbose = FALSE
)

Arguments

obj

AFMdata object

xPixel

vertical line at this pixel (1 to image resolution), if NA will use yPixel

yPixel

horizontal line at this pixel (1 to image resolution), if NA will use xPixel

no

channel number

dataOnly

if TRUE, returns data instead of AFMdata object

verbose

if TRUE, output additional information

Value

AFMdata object with line data or data frame with data

See also

Author

Thomas Gredig

Examples

filename = AFM.getSampleImages()
afmd = AFM.import(filename)
#> Error in if (grepl("ibw$", filename)) obj = read.AR_file.v2(filename) else if (grepl("tiff$",     filename)) obj = read.Park_file.v2(filename) else if (grepl("nid$",     filename)) obj = read.NanoSurf_file.v2(filename) else {    d = AFM.read(filename)    z.conv = 1    if (d$z[1] != 0)         z.conv = d$z.nm[1]/d$z[1]    d1 = list(d$z.nm)    if (is.null(attr(d, "note")))         attr(d, "note") = "none"    obj = AFMdata(data = list(z = d1), channel = attr(d, "channel"),         x.conv = max(d$x.nm)/(max(d$x) - 1), y.conv = max(d$y.nm)/(max(d$y) -             1), x.pixels = max(d$x), y.pixels = max(d$y), z.conv = z.conv,         z.units = .getChannelUnits(attr(d, "channel")), instrument = attr(d,             "instrument"), history = "", description = attr(d,             "note"), fullFilename = filename)}: the condition has length > 1
afmd2 = AFM.getLine(afmd, 50)
#> Error in AFM.getLine(afmd, 50): object 'afmd' not found
plot(afmd2, addLines = TRUE)
#> Error in plot(afmd2, addLines = TRUE): object 'afmd2' not found
head(AFM.linePlot(afmd2, dataOnly = TRUE))
#> Error in AFM.linePlot(afmd2, dataOnly = TRUE): object 'afmd2' not found